ut fix

+++ Packs/FiltersAndTransformers/Scripts/MapRangeValues/MapRangeValues_test.py
@@ -31,8 +31,8 @@ def test_rage_replace_class(start, end, value, should_replace):
 
 data_test_get_typed_value = [
     ('a', str),
-    ('1', int),
-    (1, int),
+    ('1', float),
+    (1, float),
 ]
